+++ changelog.md
@@ -5,6 +5,7 @@ TGUI 0.10.1  (TBD)
 - Added SFML/OpenGL backend (sfml-window + OpenGL >= 3.3)
 - Added GLFW/OpenGL backend (glfw3 + OpenGL >= 3.3 or GLES >= 2.0)
 - Added SDL\_Renderer backend (SDL + SDL\_ttf)
+- Added moveWithAnimation and resizeWithAnimation functions to Widget
 - TextSize can now also be set in theme file